% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pig_sentence.R
\name{pig_sentence}
\alias{pig_sentence}
\title{Translate English sentences to Pig Latin}
\usage{
pig_sentence(sentence)
}
\arguments{
\item{sentence}{A string containing the English sentence(s) to be translated to Pig Latin.}
}
\value{
A string containing the Pig Latin translation of the input sentence(s). Words are converted to lowercase and punctuation is removed from the input sentence before translation.
}
\description{
This function takes one or more English sentences as input and returns the Pig Latin translation. For the translation, this function uses the \href{https://en.wikipedia.org/wiki/Pig_Latin#Rules}{rules described in the Wikipedia article}.
}
\examples{
pig_sentence("I speak Pig Latin") ## returns "iway eakspay igpay atinlay"
}
\seealso{
[pig_word()] for translating individual words to Pig Latin.
}
