% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pig_word.R
\name{pig_word}
\alias{pig_word}
\title{Translate English words to Pig Latin}
\usage{
pig_word(word)
}
\arguments{
\item{word}{A string containing the English word to be translated to Pig Latin.}
}
\value{
A string containing the Pig Latin translation of the input word. Words are converted to lowercase before translation.
}
\description{
This function takes an English word as input and returns the Pig Latin translation of the word. For the translation, this function uses the \href{https://en.wikipedia.org/wiki/Pig_Latin#Rules}{rules described in the Wikipedia article}.
}
\examples{
pig_word("latin") ## returns "atinlay"
}
\seealso{
[pig_sentence()] for translating entire sentences to Pig Latin.
}
